+++ packages/components/Popconfirm/src/index.ts
@@ -1,0 +1,2 @@
+export { default as Popconfirm } from './Popconfirm.vue'
+export type * from './types'