+++ packages/theme/src/shortcuts/Popconfirm.ts
@@ -1,0 +1,10 @@
+export const Popconfirm = {
+  'lt-popconfirm': 'px-1.5 py-2.5',
+  'lt-popconfirm-message': 'flex items-start mb-2',
+  'lt-popconfirm-message-icon': 'w-5.5 h-5.5',
+  'lt-popconfirm-message-icon-inner': 'inline-block pt-.6 text-base',
+  'lt-popconfirm-message-text': 'text-sm',
+  'lt-popconfirm-message-text-title': 'font-600',
+  'lt-popconfirm-buttons': 'flex justify-end gap-2'
+
+}