+++ packages/theme/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @lite-space/theme
 
+## 0.4.0
+
+### Minor Changes
+
+- @lite-space/popconfirm@0.0.0
+
 ## 0.3.0
 
 ### Minor Changes